Float double checked

+++ p01_uninformed_search_c++/src/node.cc
@@ -1,7 +1,7 @@
 #include "../include/node.hpp"
 
 Node::Node(const unsigned data, const unsigned child_number, 
-           Node* predecesor, const unsigned predecesor_cost) : identifier_(data), child_number_(child_number), 
+           Node* predecesor, const double predecesor_cost) : identifier_(data), child_number_(child_number), 
                                predecesor_node_(predecesor), predecesor_cost_(predecesor_cost) {}
 
 Node::~Node() {
